refactor: add render::uniform module
