+++ reverie-engine/src/render/colored.rs
@@ -4,7 +4,7 @@ use wgpu as w;
 
 use crate::model::colored::ColoredVertex;
 
-use super::{BindingId, vertex::Vertex};
+use super::{BindingId, uniform, vertex::Vertex};
 
 pub static LOC_VERTEX: u32 = 0;
 pub static LOC_COLOR: u32 = 1;
@@ -47,8 +47,11 @@ impl ColoredRenderPipeline {
             source: w::ShaderSource::Wgsl(Cow::Borrowed(include_str!("./colored.wgsl"))),
         });
 
-        let (uniform_bind_group_layout, uniform_bind_group) =
-            create_uniform_bind_group(device, transform_uniform_buffer);
+        let (uniform_bind_group_layout, uniform_bind_group) = uniform::create_mat4_bind_group(
+            device,
+            transform_uniform_buffer,
+            BINDING_TRANSFORM.binding,
+        );
 
         let pipeline_layout = device.create_pipeline_layout(&w::PipelineLayoutDescriptor {
             label: Some("sprite model render pipeline layout"),
@@ -112,34 +115,4 @@ impl ColoredRenderPipeline {
             uniform_bind_group,
         }
     }
-}
-
-fn create_uniform_bind_group(
-    device: &w::Device,
-    transform_uniform_buffer: &w::Buffer,
-) -> (w::BindGroupLayout, w::BindGroup) {
-    let layout = device.create_bind_group_layout(&w::BindGroupLayoutDescriptor {
-        label: Some("Main Bind Group Layout"),
-        entries: &[w::BindGroupLayoutEntry {
-            binding: BINDING_TRANSFORM.binding,
-            visibility: w::ShaderStages::VERTEX,
-            ty: w::BindingType::Buffer {
-                ty: w::BufferBindingType::Uniform,
-                has_dynamic_offset: false,
-                min_binding_size: w::BufferSize::new(size_of::<[f32; 4 * 4]>() as u64),
-            },
-            count: None,
-        }],
-    });
-
-    let bind_group = device.create_bind_group(&w::BindGroupDescriptor {
-        label: Some("Main Bind Group"),
-        layout: &layout,
-        entries: &[w::BindGroupEntry {
-            binding: BINDING_TRANSFORM.binding,
-            resource: transform_uniform_buffer.as_entire_binding(),
-        }],
-    });
-
-    (layout, bind_group)
 }
